refactor splinterv2 into modules
