+++ Cargo.toml
@@ -37,3 +37,7 @@ rand = "=0.9.2"
 [[bench]]
 name = "bitmaps"
 harness = false
+
+[[bench]]
+name = "benchv2"
+harness = false
